requirements.txt
mcp[cli]
google-genai 
python-dotenv 
playwright
fastapi
uvicorn[standard]